Big update.
AI now uses a "task method". This means it creates a list of tasks which need to
be done and then assigns assets to do the tasks. This stops the AI from
assigning two colonisation fleets to the same planet.

AI still has no concept of time. It will rewrite all the orders every turn. This
is bad, but oh well.
.